api/requirements.txt
astroid==3.3.10
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
dill==0.4.0
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.1
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.13.0
pylint==3.3.7
pymongo==4.13.2
python-dotenv==1.1.0
requests==2.32.4
tomlkit==0.13.3
urllib3==1.26.20
Werkzeug==3.1.3
